Adding blocks are going to replace some others.
Blocks are saved in a object with numeric keys. The new key was
generated while incrementing the length of the keys. When the keys are
not incremented from 0 to n the new generated keys matches an other
which causes the overridden blocks in the object structure. So the new
idHelper generates the new key based on the last numeric key in the
object literal incremented by one.