update ftw.theming compatible toolbox icons.
In ftw.theming the font-awesome icon selectors have changed for fixing
an selector ordering conflict with the default icon.

Also the "icon-missing" color was changed to fuchsia.

This change fixes the extends selector so that icons work again and it
also adds an icon-missing icon in fuchsia. We really want to have icons
for all blocks ;-).