extend page-builder with .with_blocks method.
The `.with_blocks` method of the page state builder accepts one or more
block builders.

On blocks added with the `.with_blocks` method, `.within(page)` is
automatically called for placing the blocks in the page.
After the page is created, the blocks are created too and the page
configuration is synced.