Drag behaviour is not saticfactory.
When dragging to a smaller drop target the cursor does not properly
match the drop zone. So fix the width of the dragging helper and place
the cursor in the middle to handle the dirrent sizes.
To ensure that the positions are well calculated trigger a position
refresh with ``` $('.sl-column').sortable('refreshPositions')``` as seen
in https://api.jqueryui.com/sortable/#method-refreshPositions.