Fix exception when form folder is deleted.
The portal type FormFolder has been turned into a simpleyalyout block in the amnesty.ch project. Deleting such a form folder causes an exception because of a handler in ftw.simplylayout. This handler tries to adapt the parent of the form folder with IPageConfiguration. This will fail if the parent of the form folder is a folder.

See https://basecamp.com/2768704/projects/9428894/todos/238322959