Some blocks can be hidden now.
Hidden blocks will not be shown on its simplelayout container if the user does not have the permission to edit content.

Users allowed to edit the block will see the block anyway (highlighted).

This is not a security feature, the block and its content can still be accessed if the path is known.