Support migration from ftw.contentpage.
The `migration.py` provides a set of inplace migrators for migrating AT
types (`ftw.contentpage`) to DX types (`ftw.simplelayout`).

The migration must be executed in an upgrade step in the integration
package / project, because:

- a migration upgrade step can't be put in neither the old
  package (`ftw.contentpage`) nor the new package (`ftw.simplelayout`),
  because neither can be sure that the other package is present at a
  specific upgrade step (version)

- in a project the types may have had additional fields bevore (schema
  extender) and / or after (additional behaviors)

- such migrations should be tested well and implemented carefully for a
  specific project

Because of those reasons, we provide pre-configured inplace migrators
for common simplelayout types for various packages here, but they must
be executed in an upgrade step in the project code.

The inplace migrators cover the common problems when migrating
simplelayout types.