Make leadimages "has_image" and "get_scale" usable.
Before, one had to call `_get_image`, an internal method, in order to
update the view state so that `has_image` and `get_scale` would work as
expected. This is not how public methods should work.

By refactoring it this way I try to not change the public interface but
fix the state issues by auto-loading internally when not yet done.