Disable inline validation of the Plone forms in the overlay.
The main reason for this is the Plone file widget which causes some browsers to trigger the validation when the user wants to replace the file. The validation is triggered when the user clicks the button to select a file from the filesystem and he has to click it a second time for the file selector to open.