update tests to be compatible with newest testbrowser.
The testbrowser has changed:
after opening a static document (open_html) it is no longer possible to
continue with clicking in the loaded document, since the browser is
missing state such as an URL and the static driver does not support
making requests.

The correct solution is to use `.parse()` instead of `.open_html()` so
that the driver does not change and later requests are done with the
original driver.