Always get scaled image when opening colorbox overlay to fix caching issue
When an image in a galleryblock was deleted and replaced with another image
with the same name, the image would not reload in the colorbox overlay because
of image caching. By getting a scale url for each image in the overlay this is
now fixed, since it forces a reload of the image each time.