Integrate with ftw.trash.
Lets ftw.simplelayout automatically integrate with ftw.trash.

ftw.trash by default requires the "Add portal content" permission on
the parent of an object in order to restore the object.
For simplelayout blocks we want a different behavior: since we
consider blocks to be part of the content of their page, restoring
blocks should only work when the user has the "Modify portal content"
permission.
We wouldn't want users to be able to restore blocks of published
pages, the user would need to retract or revise the page first.

The integration configures ftw.simplelayout to check for the "Modify
portal content" permission whenever a user tries to restore a block.