Include map preferences behaviours in mapblock.js
This will be used for ftw.addressblock to work correctly.
The selector looks for an id starting with `field-widget` because otherwise it would also selects the widgets wrapper div. To get the zoomlevel / maplayer we filter for ending with. Together with the starting filter we can get all zoomlevel / maplayer widgets from or not from a behaviour.