Install contenttypes extras by default and deprecate the extra itself
We do this for multiple reasons:
- To fix and prevent future dependecy problems. At the moment
it was not the case that when you install ftw.simplelayout without the
[contentypes] extra the plone instance could not be installed due to
missing dependencies. To prevent this in the future and fix the current
bug we now install all the contenttypes dependencies by default.
- The [contenttypes] extra is used everywhere. We do not know of any
useage of simplelayout without the contenttypes extra. Simplelayout
without the contenttypes doesn't make much sense because the types added
is what makes simplelayout usable in the first case.