Save file as blobs
Files should be saved as blobs and not in the database as binary data.
This also fixes a bug with images that are rotated via
`exif:Orientation`. When saving images as NamedImage the rotation fix by
plone.namedfile was not applied correctly, when using NamedBlobImage
though this works like a charm.