Merge yarn and grunt watchers
We remove the default yarn watcher and reimplement it in grunt so we can
use grunt to do the work for us. yarn's watch and build are mapped to
grunt's watch and build.

To prevent the package to be released in dev mode ("grunt" was mapped to
"grunt dev" and people probably didn't notice that and the package was
released with its js in dev mode) we remove the "grunt dev" command so
we only have the default one. If you want to develop the js you can use
the option described in the gruntfile in the
requirejs section.