Fix loading error in filelistingblock.
After a migration of the filelisting block we had an issue with it to load. The sort_order is expected to be a utf-8 string but we passed in a unicode string now. This resulted in a list of Falses for every char in the unicode string instead of one False being returned in https://github.com/zopefoundation/Products.ZCatalog/blob/e751eff05374e7f8b39d59b9b9e16e33fab47e11/src/Products/ZCatalog/Catalog.py#L1085.