Fix loading new videoblock.js changes
The videoblock bundle had a compiled js set which did not exist though.
This meant that when changes in the videoblock.js were made plone did
not notice that. As a solution we load the videoblock.js together with
the rest of our javascripts. Since the contenttype are always used it
does not matter if we have it in a separate bundle or not.