Fix/implement rendering the ftw.slider block.
The ftw.slider block differs from the others. Therefore an additional condition is required. Also a fix is suggested in ftw.slider to actually add the interface prepared for the type to it. This means that the implementation in this commit will only work with the most recent code of ftw.slider.