Fix path on startpage.
On a plone site (designed as a Sl Page) a error occured when resizing an image.
The reason was that the context was wrong and therefore no site was set. When getting the block (by uuid) the page was therefore not set (the request was on application) and the object could not be found.
By adding the total path instead of only the relative part this issue can be solved.